Move longer handlers to hooks in utils
